Moved all consolidation modules into consolidation folder
Jira ticket: CAMS-405
