+++ user-interface/src/data-verification/consolidation/ConsolidationOrderAccordionView.tsx
@@ -2,7 +2,7 @@ import { Accordion } from '@/lib/components/uswds/Accordion';
 import { FormRequirementsNotice } from '@/lib/components/uswds/FormRequirementsNotice';
 import { RadioGroup } from '@/lib/components/uswds/RadioGroup';
 import Radio from '@/lib/components/uswds/Radio';
-import { ConsolidationCaseTable } from '@/data-verification/ConsolidationCasesTable';
+import { ConsolidationCaseTable } from '@/data-verification/consolidation/ConsolidationCasesTable';
 import Checkbox from '@/lib/components/uswds/Checkbox';
 import CamsSelect from '@/lib/components/CamsSelect';
 import CaseNumberInput from '@/lib/components/CaseNumberInput';
@@ -10,7 +10,7 @@ import Alert, { UswdsAlertStyle } from '@/lib/components/uswds/Alert';
 import { LoadingSpinner } from '@/lib/components/LoadingSpinner';
 import { CaseTable } from '@/data-verification/transfer/CaseTable';
 import Button, { UswdsButtonStyle } from '@/lib/components/uswds/Button';
-import { ConsolidationOrderModal } from '@/data-verification/ConsolidationOrderModal';
+import { ConsolidationOrderModal } from '@/data-verification/consolidation/ConsolidationOrderModal';
 import { CaseNumber } from '@/lib/components/CaseNumber';
 import { ConsolidationViewModel } from '@/data-verification/consolidation/consolidationViewModel';
 import { getCaseNumber } from '@/lib/utils/formatCaseNumber';
